requirements.txt
opencv-python>=3.6.2
numpy>=1.13.3
python-constraint>=1.4.0
